Database: New module for SQLite Database interaction
Adds a new Database module with the ability to open connections and run
queries with a SQLite databases.

We can open a connection with a SQLite database, close a connection, run
queries in a couple of different ways, and get the id of the last row
inserted into the Database.

Documentation and tests for the feature have been added as well.